Changed to use solely junit 5 syntax, refactored PCAP.py, moved to nifi-network-bundle, renamed to follow acronym naming conventions, made other miscellaneous review changes
