+++ nifi-extension-bundles/nifi-network-bundle/nifi-network-processors/src/test/java/org/apache/nifi/processors/network/util/TestPCAP.java
@@ -16,27 +16,20 @@
  */
 
 
-package org.apache.nifi.processors.standard.util;
+package org.apache.nifi.processors.network.util;
 
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.util.ArrayList;
 
-import org.apache.nifi.processors.standard.util.Pcap.Packet;
-import org.apache.nifi.processors.standard.util.Pcap.Header;
+import org.apache.nifi.processors.network.util.PCAP.Packet;
+import org.apache.nifi.processors.network.util.PCAP.Header;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import org.junit.Assert;
+import org.junit.jupiter.api.Assertions;
 
-public class TestPcap {
-
-    @BeforeEach
-    public void init() {
-
-    }
-
+public class TestPCAP {
     @Test
     public void testReadBytesFull() {
 
@@ -67,34 +60,32 @@ public class TestPcap {
 
         // create test PCAP
 
-        Pcap testPcap = new Pcap(hdr, packets);
+        PCAP testPcap = new PCAP(hdr, packets);
 
         // Call the readBytesFull method
         byte[] result = testPcap.readBytesFull();
 
         // Assert the expected byte array length
-        Assert.assertEquals(70, result.length);
+        Assertions.assertEquals(70, result.length);
 
         // Assert the expected byte array values
         ByteBuffer buffer = ByteBuffer.wrap(result);//.order(ByteOrder.LITTLE_ENDIAN);
-        Assert.assertEquals(0xa1b2c3d4, buffer.getInt());
+        Assertions.assertEquals(0xa1b2c3d4, buffer.getInt());
         ByteBuffer LEBuffer = ByteBuffer.wrap(result).order(ByteOrder.LITTLE_ENDIAN);
         LEBuffer.position(4);
-        Assert.assertEquals(2, LEBuffer.getShort());
-        Assert.assertEquals(4, LEBuffer.getShort());
-        Assert.assertEquals(0, LEBuffer.getInt());
-        Assert.assertEquals(0, LEBuffer.getInt());
-        Assert.assertEquals(40, LEBuffer.getInt());
-        Assert.assertEquals(1, LEBuffer.getInt());
-
-
-        Assert.assertEquals(1713184965, LEBuffer.getInt());
-        Assert.assertEquals(1000, LEBuffer.getInt());
-        Assert.assertEquals(30, LEBuffer.getInt());
-        Assert.assertEquals(30, LEBuffer.getInt());
+        Assertions.assertEquals(2, LEBuffer.getShort());
+        Assertions.assertEquals(4, LEBuffer.getShort());
+        Assertions.assertEquals(0, LEBuffer.getInt());
+        Assertions.assertEquals(0, LEBuffer.getInt());
+        Assertions.assertEquals(40, LEBuffer.getInt());
+        Assertions.assertEquals(1, LEBuffer.getInt());
+        Assertions.assertEquals(1713184965, LEBuffer.getInt());
+        Assertions.assertEquals(1000, LEBuffer.getInt());
+        Assertions.assertEquals(30, LEBuffer.getInt());
+        Assertions.assertEquals(30, LEBuffer.getInt());
         byte[] bodyArray = new byte[30];
         LEBuffer.get(40, bodyArray, 0, 30).array();
-        Assert.assertArrayEquals(new byte[]{
+        Assertions.assertArrayEquals(new byte[]{
             0,  1,  2,  3,  4,  5,  6,  7,  8,  9,
             10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
             20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
